All good

+++ backend/booking-service/src/main/java/org/example/dto/BookingResponse.java
@@ -1,16 +1,16 @@
 package org.example.dto;
 
-import org.example.domain.BookingStatus;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
+import org.example.domain.BookingStatus;
+
 public record BookingResponse(
-  Long id,
-  Long userId,
-  Long carId,
-  LocalDateTime startDate,
-  LocalDateTime endDate,
-  BookingStatus status,
-  Float totalPrice,
-  LocalDateTime createdAt
-) {}
+        Long id,
+        Long userId,
+        Long carId,
+        LocalDateTime startDate,
+        LocalDateTime endDate,
+        BookingStatus status,
+        BigDecimal totalPrice,
+        LocalDateTime createdAt) {}
